Optimise Dockerfile to maximise layer reuse and minimise build time
First, run pip install when only the setup.py has been copied in to install
dependencies. Doing this first means dependencies (which take time to
install, and result in a relatively large layer) are only re-installed if
the setup.py file changes. Of course, having the version number in setup.py
means it will change every new version, but this will help during dev.

Then, copy code in and run pip install again to actually install the
exporter.

Finally, copy in license and readme. Changes to these won't trigger the
more expensive layers, and these are so cheap it doesn't matter if we run
them all the time.